move deck_view_container to src/game/deckview
