+++ cockatrice/src/client/tabs/tab_game.cpp
@@ -6,6 +6,7 @@
 #include "../../game/cards/card_database.h"
 #include "../../game/cards/card_database_manager.h"
 #include "../../game/cards/card_item.h"
+#include "../../game/deckview/deck_view_container.h"
 #include "../../game/game_scene.h"
 #include "../../game/game_view.h"
 #include "../../game/player/player.h"
@@ -21,7 +22,6 @@
 #include "../ui/phases_toolbar.h"
 #include "../ui/picture_loader/picture_loader.h"
 #include "../ui/window_main.h"
-#include "./deck_view_container.h"
 #include "get_pb_extension.h"
 #include "pb/command_concede.pb.h"
 #include "pb/command_delete_arrow.pb.h"
