+++ cockatrice/src/game/deckview/deck_view_container.cpp
@@ -1,23 +1,21 @@
 #include "deck_view_container.h"
 
+#include "../../client/tabs/tab_game.h"
+#include "../../client/ui/picture_loader/picture_loader.h"
 #include "../../deck/deck_loader.h"
 #include "../../dialogs/dlg_load_deck.h"
 #include "../../dialogs/dlg_load_remote_deck.h"
-#include "../../game/cards/card_database.h"
-#include "../../game/cards/card_database_manager.h"
-#include "../../game/deckview/deck_view.h"
-#include "../../game/game_scene.h"
-#include "../../game/game_view.h"
 #include "../../server/pending_command.h"
 #include "../../settings/cache_settings.h"
-#include "../game_logic/abstract_client.h"
-#include "../ui/picture_loader.h"
+#include "../cards/card_database.h"
+#include "../cards/card_database_manager.h"
+#include "../game_scene.h"
+#include "deck_view.h"
 #include "pb/command_deck_select.pb.h"
 #include "pb/command_ready_start.pb.h"
 #include "pb/command_set_sideboard_lock.pb.h"
 #include "pb/command_set_sideboard_plan.pb.h"
 #include "pb/response_deck_download.pb.h"
-#include "tab_game.h"
 #include "trice_limits.h"
 
 #include <QMessageBox>
