+++ cockatrice/src/game/deckview/deck_view_container.h
